make clang happy

+++ src/utils.cpp
@@ -6,7 +6,7 @@ float randomRange( float a, float b )
 {
     thread_local std::mt19937 random( std::random_device{}() );
     const uint64_t r = random();
-    return ( b - a ) * static_cast<float>( r ) / random.max() + a;
+    return ( b - a ) * static_cast<float>( r ) / (float)random.max() + a;
 }
 
 double colorHalf( double col )
